refactor(review): 拆分 train_program_page.dart (767行) — 详情页抽到 part 文件
- train_program_page.dart：仅保留列表页（筛选/学院年级下拉/方案列表），269 行
- train_program_detail_page.dart：详情页（方案概览 + 课程架构树 + 课程详情
  bottom sheet），约 500 行
- 私有 State 类经 part 共享主 library import，外部零改动
